polish(nyz): add cloudpickle clone and fix minor bugs
