+++ ding/utils/__init__.py
@@ -1,6 +1,6 @@
 import ding
 from .collection_helper import iter_mapping
-from .compression_helper import get_data_compressor, get_data_decompressor
+from .compression_helper import get_data_compressor, get_data_decompressor, CloudPickleWrapper
 from .default_helper import override, dicts_to_lists, lists_to_dicts, squeeze, default_get, error_wrapper, list_split, \
     LimitedSpaceContainer, deep_merge_dicts, set_pkg_seed, flatten_dict, one_time_warning, split_data_generator, \
     RunningMeanStd, make_key_as_identifier, remove_illegal_item
